+++ test/files/rules/nounusedvariable-imports.test.ts
@@ -1,5 +1,6 @@
 import  $ = require("jquery");
 import  _ = require("underscore");
 import  xyz = require("xyz");
+export import a = require("a");
 
 $(_.xyz());